+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 ## 0.3.0 (Unreleased)
 
 - Prevent _test_test.go PR file lookup ([#20](https://github.com/katbyte/tctest/issues/20))
+- Rename `status` to `results` for accuracy ([#15](https://github.com/katbyte/tctest/issues/15))
+- Rename root command to `branch` to avoid triggering accidental builds on typos ([#15](https://github.com/katbyte/tctest/issues/15))
+- Add `--wait` option to `pr`, `branch`, and `results` commands ([#15](https://github.com/katbyte/tctest/issues/15))
+- `list` command no longer triggers a build when no tests are found ([#15](https://github.com/katbyte/tctest/issues/15))
+- Usage information is no longer displayed after non-usage-related errors ([#15](https://github.com/katbyte/tctest/issues/15))
+- `results` command will display a warning if an in-progress build would give incomplete results ([#15](https://github.com/katbyte/tctest/issues/15))
+- `results` will now inform the user if the specified build is still queued ([#15](https://github.com/katbyte/tctest/issues/15))
 
 ## 0.2.0 (Jan 22th, 2020)